requirements.txt
pandas
sqlalchemy
python_dotenv
clickhouse_driver
clickhouse_sqlalchemy